+++ src/core/NamedObjs.cpp
@@ -9,7 +9,7 @@
 
 __removed #pragma package(smart_init)
 
-static intptr_t NamedObjectSortProc(const void *Item1, const void *Item2)
+static int32_t NamedObjectSortProc(const void *Item1, const void *Item2)
 {
   return get_as<TNamedObject>(Item1)->Compare(get_as<TNamedObject>(Item2));
 }
@@ -26,9 +26,9 @@ void TNamedObject::SetName(const UnicodeString Value)
   FName = Value;
 }
 
-intptr_t TNamedObject::Compare(const TNamedObject *Other) const
+int32_t TNamedObject::Compare(const TNamedObject *Other) const
 {
-  intptr_t Result;
+  int32_t Result;
   if (GetHidden() && !Other->GetHidden())
   {
     Result = -1;
@@ -57,7 +57,7 @@ void TNamedObject::MakeUniqueIn(TNamedObjectList *List)
     while (List->FindByName(GetName()))
     {
       int64_t N = 0;
-      intptr_t P;
+      int32_t P;
       // If name already contains number parenthesis remove it (and remember it)
       UnicodeString Name = GetName();
       if ((Name[Name.Length()] == L')') && ((P = Name.LastDelimiter(L'(')) > 0))
@@ -85,19 +85,19 @@ TNamedObjectList::TNamedObjectList(TObjectClassId Kind) noexcept :
 {
 }
 
-const TNamedObject * TNamedObjectList::AtObject(intptr_t Index) const
+const TNamedObject * TNamedObjectList::AtObject(int32_t Index) const
 {
   return const_cast<TNamedObjectList *>(this)->AtObject(Index);
 }
 
-TNamedObject * TNamedObjectList::AtObject(intptr_t Index)
+TNamedObject * TNamedObjectList::AtObject(int32_t Index)
 {
   return GetAs<TNamedObject>(Index + FHiddenCount);
 }
 
 void TNamedObjectList::Recount()
 {
-  intptr_t Index = 0;
+  int32_t Index = 0;
   while ((Index < TObjectList::GetCount()) && GetAs<TNamedObject>(Index)->GetHidden())
   {
     ++Index;
@@ -111,9 +111,9 @@ void TNamedObjectList::AlphaSort()
   Recount();
 }
 
-intptr_t TNamedObjectList::Add(TObject *AObject)
+int32_t TNamedObjectList::Add(TObject *AObject)
 {
-  intptr_t Result;
+  int32_t Result;
   TAutoFlag ControlledAddFlag(FControlledAdd); nb::used(ControlledAddFlag);
   TNamedObject *NamedObject = static_cast<TNamedObject *>(AObject);
   // If temporarily not auto-sorting (when loading session list),
@@ -175,18 +175,18 @@ TNamedObject * TNamedObjectList::FindByName(const UnicodeString AName)
   return nullptr;
 }
 
-void TNamedObjectList::SetCount(intptr_t Value)
+void TNamedObjectList::SetCount(int32_t Value)
 {
   TObjectList::SetCount(Value/*+HiddenCount*/);
 }
 
-intptr_t TNamedObjectList::GetCount() const
+int32_t TNamedObjectList::GetCount() const
 {
   DebugAssert(FHiddenCount >= 0);
   return TObjectList::GetCount() - FHiddenCount;
 }
 
-intptr_t TNamedObjectList::GetCountIncludingHidden() const
+int32_t TNamedObjectList::GetCountIncludingHidden() const
 {
   DebugAssert(FHiddenCount >= 0);
   return TObjectList::GetCount();
